Uppdaterat nya klasser utifrån Fö 4

+++ include/ColorDBL.h
@@ -5,28 +5,28 @@
 // Behöver kanske ändra så färger är mellan 0 och 1 för att det ska funka med programeringen
 class ColorDBL {
 private:
-	double R = 0.0;
-	double G = 0.0;
-	double B = 0.0;
+	float R = 0.0f;
+	float G = 0.0f;
+	float B = 0.0f;
 
 public:
 	ColorDBL() {};
 
-	ColorDBL(const double& r, const double& g, const double& b): R{ std::min(std::max(r, 0.0), 1.0) },
-											G{ std::min(std::max(g, 0.0), 1.0) }, 
-											B{ std::min(std::max(b, 0.0), 1.0) } {}
+	ColorDBL(const float& r, const float& g, const float& b): R{ std::min(std::max(r, 0.0f), 1.0f) },
+											G{ std::min(std::max(g, 0.0f), 1.0f) }, 
+											B{ std::min(std::max(b, 0.0f), 1.0f) } {}
 
-	double getR() { return R; }
+	float getR() { return R; }
 
-	double getG() { return G; }
+	float getG() { return G; }
 
-	double getB() { return R; }
+	float getB() { return R; }
 
-	void setR(const double& r) { R = r; }
+	void setR(const float& r) { R = r; }
 
-	void setG(const double& g) { G = g; }
+	void setG(const float& g) { G = g; }
 
-	void setB(const double& b) { B = b; }
+	void setB(const float& b) { B = b; }
 
 	friend std::ostream& operator<<(std::ostream& os, const ColorDBL& c) {
 		os << "[ " << c.R * 255.0 << ", " << c.G * 255 << ", " << c.B * 255.0 << " ] / 255\n";
